adding controllers, put place funcitons in own controller, comments on place beginning implementation
